refactor: update Python path in launch configuration and enhance module imports
- Adjusted the Python path in the VSCode launch configuration to point to the correct environment.
- Added project root to `sys.path` in `run.py` and `__init__.py` to ensure shared packages are accessible.
- Cleaned up imports and refactored internal token verification logic in the `auth` routes.
- Introduced shared schemas in `backend_shared` for authentication and internal requests.
- Removed unused code and improved error handling in WebSocket routes.
