+++ .vscode/launch.json
@@ -20,7 +20,7 @@
             "program": "${workspaceFolder}/backend/run.py",
             "console": "integratedTerminal",
             "envFile": "${workspaceFolder}/backend/.env",
-            "pythonPath": "${workspaceFolder}/backend/genassist_env/bin/python",
+            "pythonPath": "${workspaceFolder}/genassist_env/bin/python",
             "justMyCode": true
         },
         {
@@ -31,7 +31,7 @@
             "program": "${workspaceFolder}/websocket/run.py",
             "console": "integratedTerminal",
             "envFile": "${workspaceFolder}/websocket/.env",
-            "pythonPath": "${workspaceFolder}/backend/genassist_env/bin/python",
+            "pythonPath": "${workspaceFolder}/genassist_env/bin/python",
             "justMyCode": true
         }
     ]
