+++ backend_shared/schemas/internal.py
@@ -1,0 +1,8 @@
+
+
+from pydantic import BaseModel
+
+class AgentExecuteRequest(BaseModel):
+    agent_id: str
+    thread_id: str
+    text: str
